更新了部分后端代码以及前端代码，添加了算法模型的简单模拟

+++ JavaCode/src/main/java/com/chan/controller/ApiController.java
@@ -1,16 +1,19 @@
 package com.chan.controller;
 
+import com.alibaba.fastjson2.JSON;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import com.chan.entity.CadContent;
-import com.chan.entity.Data;
-import com.chan.service.impl.DataServiceImpl;
+import com.chan.entity.CadData;
+import com.chan.service.impl.CadDataServiceImpl;
 import com.chan.service.impl.GrpcScriptExecImpl;
 import jakarta.annotation.Resource;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
 
+@CrossOrigin
 @RequestMapping("/CadDet")
 @RestController
 public class ApiController {
@@ -19,7 +22,7 @@ public class ApiController {
      * 2. 数据库CRUD操作
      */
 
-    private Data tempData; // 存储py脚本执行的数据（临时）
+    private Map<String, Object> tempDataMap;
 
     private GrpcScriptExecImpl grpcScriptExec;
 
@@ -31,31 +34,38 @@ public class ApiController {
     // 调用py接口
     @PostMapping("CallPy")
     public Object script(@RequestBody Map<String, Object> request) {
+
         String path = (String) request.getOrDefault("path", "");
         if (path.isEmpty()) {
             return "";
         }
 
-        Data data = new Data(this.grpcScriptExec.exec(new CadContent(path))
-                .getResult());
-        this.tempData = data;
-        return data.getContent();
+        String result = this.grpcScriptExec.exec(new CadContent(path)).getResult();
+        tempDataMap = JSON.parseObject(result);
+
+        return result;
     }
 
 
     @Resource
-    private DataServiceImpl dataService;
+    private CadDataServiceImpl dataService;
 
     // 数据库CRUD接口
     @PostMapping("/create")
-    public String create() {
-        if (this.tempData != null) {
-            dataService.save(this.tempData);
-            tempData = null;
+    public Object create() {
+
+        if (this.tempDataMap != null) {
+            for (String obj : tempDataMap.keySet()) {
+                Object value = tempDataMap.get(obj);
+
+                dataService.save(new CadData(obj, value.toString()));
+                System.out.println("Saving to database: Key = " + obj + ", Value = " + value);
+            }
+            tempDataMap = null;
         } else {
-            return "结果为空或未执行算法";
+            return ResponseEntity.badRequest().body("结果为空或未执行算法");
         }
-        return "create success";
+        return ResponseEntity.ok("create success");
     }
 
     @GetMapping("/read")
@@ -64,18 +74,36 @@ public class ApiController {
     }
 
     @PutMapping("/update")
-    public Object update(int id, String content) {
-        UpdateWrapper<Data> updateWrapper = new UpdateWrapper<>();
-        updateWrapper.eq("id", id);
-        updateWrapper.set("content", content);
-        dataService.update(updateWrapper);
-        return "update success";
+    public Object update(@RequestParam Integer id, String position, String value) {
+
+        if (id == null || position == null || position.isEmpty() || value == null || value.isEmpty()) {
+        return ResponseEntity.badRequest().body("参数不能为空");
+    }
+        UpdateWrapper<CadData> updateWrapper = new UpdateWrapper<>();
+        updateWrapper.eq("id", id)
+                .set("position", position)
+                .set("value", value);
+        boolean isUpdate = dataService.update(updateWrapper);
+        if (isUpdate) {
+            return ResponseEntity.ok("数据更新成功");
+        } else {
+            return ResponseEntity.badRequest().body("数据更新失败，请检查id是否存在");
+        }
     }
 
     @DeleteMapping("/delete")
-    public String delete(int id) {
-        dataService.removeById(id);
-        return "delete success";
+    public Object delete(@RequestParam Integer id) {
+
+        if (id == null) {
+            return ResponseEntity.badRequest().body("id参数不能为空");
+        }
+        boolean isDelete = dataService.removeById(id);
+        if (isDelete) {
+            return ResponseEntity.ok("数据删除成功");
+        } else {
+            return ResponseEntity.badRequest().body("数据删除失败");
+        }
+
     }
 
 }
